refactor(watch-provider): typo fix
